\name{Beetle_distM}
\alias{Beetle_distM}
\docType{data}
\title{
  Species pairwise distance matrix for beetle species
}
\description{
  This dataset includes a beetle species pairwise symmetric distance matrix for 1308 species listed in the \code{Beetle_abundance_data} dataset. Each element in the matrix represents a Gower distance between two species computed from species traits. \cr
}
\usage{data("Beetle_distM")}
\format{
  \code{Beetle_distM} is a 1308 x 1308 (species by species) symmetric data.frame. Each element of the data.frame is between zero (for species pairs with identical traits) and one. \cr
}

\keyword{datasets}
