\name{Beetle_tree}
\alias{Beetle_tree}
\docType{data}
\title{
  Phylogenetic tree for beetle abundance data
}
\description{
  This dataset includes a phylogenetic tree spanned by 1308 species listed in the \code{Beetle_abundance_data} dataset. \cr
}
\usage{data("Beetle_tree")}
\format{
  \code{Beetle_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list of 4: \cr
  $ edge       : int [1:1720, 1:2] 1309 1310 1311 1312 1313 1310 1311 ... \cr
  $ edge.length: num [1:1720] 0.1091 0.4511 0.2800 0.0349 0.1000 0.0160 ... \cr
  $ Nnode      : int 413 \cr
  $ tip.label  : chr [1:1308] "Otiorhynchus_porcatus" "Otiorhynchus_crataegi" "Otiorhynchus_pupillatus" ... \cr
}

\keyword{datasets}
