\name{Bird_incidence_data}
\alias{Bird_incidence_data}
\docType{data}
\title{
  Bird incidence raw data (datatype = "incidence_raw" for TD, PD, and FD)
}
\description{
  This dataset includes bird species incidence data  in eleven sites (B04, B05, B06, B07, H09, L11, P08, S10, U01, U02, U03) with two treatments (Control and Enhanced). In each combinations of eleven sites and two treatments (B04C, B04E, B05C, B05E, ...), there are 84, 84, 49, 69, 75, 84, 71, 57, 62, 77, 48, 78, 84, 84, 64, 83, 77, 69, 84, 55, 84, 51 sampling units.
}
\usage{data("Bird_incidence_data")}
\format{
  \code{Bird_incidence_data} is a matrix where:
  \itemize{
    \item The first three columns represent Site, Treatment, and Patch.
    \item The remaining 72 columns correspond to species names.
  }
}


\keyword{datasets}
