\name{Bird_tree}
\alias{Bird_tree}
\docType{data}
\title{
  Phylogenetic tree for bird incidence data
}
\description{
  This dataset includes a phylogenetic tree spanned by 72 species listed in the \code{Bird_incidence_data} dataset. \cr
}
\usage{data("Bird_tree")}
\format{
  \code{Bird_tree} is a list (phylo tree) with the following phylogenetic information: \cr
  A list with 4 elements: \cr
  $ edge       : int [1:135, 1:2] 73 74 75 75 74 73 1 2 3 76 ... \cr
  $ edge.length: num [1:30] 19.1 68.5 15.8 84.3 16.5 ... \cr
  $ Nnode      : int 64 \cr
  $ tip.label  : chr [1:16] "Bonasa_bonasia" "Tetrao_urogallus" "Bucephala_clangula" ... \cr
}

\keyword{datasets}
